PROEV-290 ANDROID приложение вылетает если перейти на экран регистрации с недоступной сетью (предположительно подсказки email не отлавливают ошибки)
