+++ app/src/main/java/ru/myproevent/ui/fragments/authorization/RegistrationFragment.kt
@@ -29,16 +29,10 @@ import ru.myproevent.ui.presenters.authorization.registration.RegistrationPresen
 import ru.myproevent.ui.presenters.authorization.registration.RegistrationView
 import ru.myproevent.ui.presenters.main.RouterProvider
 import java.lang.reflect.Field
-import android.widget.ArrayAdapter
-import com.jakewharton.rxbinding2.widget.RxTextView
-import java.util.concurrent.TimeUnit
 
 class RegistrationFragment :
     BaseMvpFragment<FragmentRegistrationBinding>(FragmentRegistrationBinding::inflate),
     RegistrationView {
-
-    var adapter: ArrayAdapter<String>? = null
-
     // TODO: вынести в кастомную вьюху
     private val licenceTouchListener = View.OnTouchListener { v, event ->
         when (event.action) {
@@ -96,16 +90,7 @@ class RegistrationFragment :
         emailEdit.doAfterTextChanged {
             presenter.emailEdited()
         }
-        adapter =
-            ArrayAdapter(requireContext(), android.R.layout.simple_dropdown_item_1line, arrayOf())
-        binding.emailEdit.setAdapter(adapter)
-
-        RxTextView.textChangeEvents(emailEdit)
-            .debounce(500, TimeUnit.MILLISECONDS)
-            .subscribe {
-                presenter.typedEmail(it.text().toString())
-            }
-
+        binding.emailEdit.initHints { presenter.typedEmail(it) }
 
         passwordEdit.doAfterTextChanged {
             presenter.passwordEdited()
@@ -236,8 +221,6 @@ class RegistrationFragment :
         )
 
     override fun setEmailHint(emailSuggestion: List<Suggestion>) {
-        adapter?.clear()
-        adapter?.addAll(emailSuggestion.map { it.value })
-        adapter?.filter?.filter(binding.emailEdit.text, null);
+        binding.emailEdit.setEmailHint(emailSuggestion)
     }
 }
