+++ app/src/main/java/ru/myproevent/ui/presenters/authorization/registration/RegistrationPresenter.kt
@@ -1,9 +1,12 @@
 package ru.myproevent.ui.presenters.authorization.registration
 
 import android.annotation.SuppressLint
+import android.util.Log
 import com.github.terrakok.cicerone.Router
 import io.reactivex.disposables.Disposable
 import io.reactivex.observers.DisposableCompletableObserver
+import retrofit2.HttpException
+import ru.myproevent.domain.models.repositories.email_hint.IEmailHintRepository
 import ru.myproevent.domain.models.repositories.internet_access_info.IInternetAccessInfoRepository
 import ru.myproevent.domain.models.repositories.proevent_login.IProEventLoginRepository
 import ru.myproevent.ui.presenters.BaseMvpPresenter
@@ -41,6 +44,9 @@ class RegistrationPresenter(localRouter: Router) : BaseMvpPresenter<Registration
 
     @Inject
     lateinit var loginRepository: IProEventLoginRepository
+
+    @Inject
+    lateinit var emailHintRepository: IEmailHintRepository
 
     @Inject
     lateinit var interAccessInfoRepository: IInternetAccessInfoRepository
@@ -100,9 +106,11 @@ class RegistrationPresenter(localRouter: Router) : BaseMvpPresenter<Registration
     }
 
     fun typedEmail(partEmail: String) {
-        loginRepository.getEmailHint(partEmail)
+        emailHintRepository.getEmailHint(partEmail)
             .observeOn(uiScheduler)
-            .subscribe { it -> viewState.setEmailHint(it) }
-            .disposeOnDestroy()
+            .subscribe(
+                { viewState.setEmailHint(it) },
+                { Log.e("EMAIL_HINT", it.localizedMessage ?: "") }
+            ).disposeOnDestroy()
     }
 }
